Fixing compile errors and warnings on linux build targets.

+++ test/src/test_multi_mip_error_handling.c
@@ -9,12 +9,12 @@ typedef struct {
     FILE *fp;
 } file_istream_data_t;
 
-static int64_t file_istream_tell(oapvd_istream_t *stream) {
+static long long file_istream_tell(oapvd_istream_t *stream) {
     file_istream_data_t *data = (file_istream_data_t*)stream->data;
     return oapv_ftell(data->fp);
 }
 
-static int file_istream_seek(oapvd_istream_t *stream, int64_t offset, int origin) {
+static int file_istream_seek(oapvd_istream_t *stream, long long offset, int origin) {
     file_istream_data_t *data = (file_istream_data_t*)stream->data;
     return oapv_fseek(data->fp, offset, origin);
 }
